Unrefactor informerfactory
