+++ cmd/nsdns/watch.go
@@ -10,6 +10,9 @@ import (
 import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
+	apinetworkingv1 "k8s.io/api/networking/v1"
+	"k8s.io/client-go/informers"
+	"k8s.io/client-go/tools/cache"
 )
 
 import (
@@ -58,7 +61,33 @@ func watchCommand() *cobra.Command {
 				}
 			}()
 
-			informerFactory := DomainManagerInformerFactory(dm, clientset)
+			informerFactory := informers.NewSharedInformerFactory(clientset, time.Minute)
+
+			informerFactory.Networking().V1().Ingresses().Informer().AddEventHandler(
+				cache.ResourceEventHandlerFuncs{
+					AddFunc: func(obj interface{}) {
+						ingress := obj.(*apinetworkingv1.Ingress)
+
+						if err := dm.HandleIngressExists(ingress); err != nil {
+							log.Error(err)
+						}
+					},
+					DeleteFunc: func(obj interface{}) {
+						ingress := obj.(*apinetworkingv1.Ingress)
+
+						if err := dm.HandleIngressDeleted(ingress); err != nil {
+							log.Error(err)
+						}
+					},
+					UpdateFunc: func(old, new interface{}) {
+						ingress := new.(*apinetworkingv1.Ingress)
+
+						if err := dm.HandleIngressExists(ingress); err != nil {
+							log.Error(err)
+						}
+					},
+				},
+			)
 
 			<-done
